+++ features/admin.verifiable-credentials.v1/components/vc-template-offer.tsx
@@ -21,8 +21,7 @@ import { AlertLevels, IdentifiableComponentInterface } from "@wso2is/core/models
 import { addAlert } from "@wso2is/core/store";
 import {
     ConfirmationModal,
-    CopyInputField,
-    EmphasizedSegment
+    CopyInputField
 } from "@wso2is/react-components";
 import { AxiosError } from "axios";
 import React, { FunctionComponent, ReactElement, useState } from "react";
@@ -116,7 +115,7 @@ export const VCTemplateOffer: FunctionComponent<VCTemplateOfferPropsInterface> =
     };
 
     return (
-        <EmphasizedSegment padded="very">
+        <>
             <Grid>
 
                 <Grid.Row>
@@ -141,6 +140,7 @@ export const VCTemplateOffer: FunctionComponent<VCTemplateOfferPropsInterface> =
                                                     <Button
                                                         color="red"
                                                         className="fluid"
+                                                        type="button"
                                                         loading={ isGenerating }
                                                         disabled={ isGenerating || readOnly }
                                                         onClick={ () => setShowRegenerateConfirmation(true) }
@@ -161,6 +161,7 @@ export const VCTemplateOffer: FunctionComponent<VCTemplateOfferPropsInterface> =
                                         </Message>
                                         <Button
                                             primary
+                                            type="button"
                                             loading={ isGenerating }
                                             disabled={ isGenerating || readOnly }
                                             onClick={ handleGenerateOffer }
@@ -202,6 +203,6 @@ export const VCTemplateOffer: FunctionComponent<VCTemplateOfferPropsInterface> =
                     </ConfirmationModal>
                 )
             }
-        </EmphasizedSegment>
+        </>
     );
 };