+++ features/admin.login-flow-builder.v1/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @wso2is/admin.login-flow-builder.v1
 
+## 2.29.0
+
+### Minor Changes
+
+- [#9577](https://github.com/wso2/identity-apps/pull/9577) [`5e96563efa7eb0807db57487b417a2f820474e36`](https://github.com/wso2/identity-apps/commit/5e96563efa7eb0807db57487b417a2f820474e36) Thanks [@NipuniBhagya](https://github.com/NipuniBhagya)! - Add visual editor notification system and improve api based login flow configuration
+
+### Patch Changes
+
+- Updated dependencies [[`5e96563efa7eb0807db57487b417a2f820474e36`](https://github.com/wso2/identity-apps/commit/5e96563efa7eb0807db57487b417a2f820474e36)]:
+  - @wso2is/admin.applications.v1@2.41.0
+  - @wso2is/admin.core.v1@2.50.5
+  - @wso2is/admin.server-configurations.v1@2.36.5
+
 ## 2.28.2
 
 ### Patch Changes